+++ app/src/main/java/com/gitsurfer/gitsurf/data/network/models/response/Payload.kt
@@ -1,6 +1,6 @@
 package com.gitsurfer.gitsurf.data.network.models.response
 
-import com.gitsurfer.gitsurf.data.roomdatabase.models.RoomPayload
+import com.gitsurfer.gitsurf.data.persistence.models.RoomPayload
 import com.squareup.moshi.Json
 
 data class Payload(
