+++ app/src/main/java/com/gitsurfer/gitsurf/data/persistence/base/BaseDAO.kt
@@ -1,4 +1,4 @@
-package com.gitsurfer.gitsurf.data.roomdatabase.base
+package com.gitsurfer.gitsurf.data.persistence.base
 
 import androidx.room.Delete
 import androidx.room.Insert
