--- app/src/main/java/com/gitsurfer/gitsurf/data/usecases/base/SuspendedPerformUseCase.kt
@@ -1,9 +1,0 @@
-package com.gitsurfer.gitsurf.data.usecases.base
-
-/**
- * Base use case that does not take any input parameters and returns results
- */
-
-abstract class SuspendedPerformUseCase<out Results> {
-  abstract suspend fun perform(): Results
-}
