--- app/src/main/java/com/gitsurfer/gitsurf/data/usecases/base/SuspendedPrepareUseCase.kt
@@ -1,10 +1,0 @@
-package com.gitsurfer.gitsurf.data.usecases.base
-
-/**
- * Base use case that takes Prepare parameters, executable parameters and returns result
- */
-abstract class SuspendedPrepareUseCase<out Results, in PrepareParams, in ExecutableParam> :
-    SuspendedUseCase<Results, ExecutableParam>() {
-
-  abstract suspend fun prepare(params: PrepareParams)
-}
