--- app/src/main/java/com/gitsurfer/gitsurf/injection/scope/ActivityScope.kt
@@ -1,6 +1,0 @@
-package com.gitsurfer.gitsurf.injection.scope
-
-import javax.inject.Scope
-
-@Scope
-annotation class ActivityScope
