--- app/src/main/java/com/gitsurfer/gitsurf/injection/scope/FragmentScope.kt
@@ -1,14 +1,0 @@
-package com.gitsurfer.gitsurf.injection.scope
-
-import javax.inject.Scope
-
-@Scope
-@kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
-@Target(
-    AnnotationTarget.CLASS,
-    AnnotationTarget.FILE,
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.PROPERTY_SETTER
-)
-annotation class FragmentScope
