+++ app/src/main/java/com/gitsurfer/gitsurf/utils/exceptions/ValidationException.kt
@@ -1,4 +1,4 @@
-package com.gitsurfer.gitsurf.data.usecases.exceptions
+package com.gitsurfer.gitsurf.utils.exceptions
 
 sealed class ValidationException : Exception() {
   object UsernameEmpty : ValidationException()
