+++ app/src/main/java/com/gitsurfer/gitsurf/utils/ui/SnackBarAction.kt
@@ -1,4 +1,4 @@
-package com.gitsurfer.gitsurf.utils
+package com.gitsurfer.gitsurf.utils.ui
 
 import android.view.View
