+++ app/src/main/res/layout/item_room_feed.xml
@@ -8,7 +8,7 @@
 
     <variable
         name="roomFeed"
-        type="com.gitsurfer.gitsurf.data.roomdatabase.models.RoomFeed"
+        type="com.gitsurfer.gitsurf.data.persistence.models.RoomFeed"
         />
     <variable
         name="HtmlCompat"
