+++ app/src/main/res/layout/nav_header.xml
@@ -7,7 +7,7 @@
   <data>
     <variable
         name="roomUser"
-        type="com.gitsurfer.gitsurf.data.roomdatabase.models.RoomUser"
+        type="com.gitsurfer.gitsurf.data.persistence.models.RoomUser"
         />
   </data>
